Add .webp support

+++ modules/pdf.go
@@ -44,12 +44,12 @@ func PDF(s *discordgo.Session, m *discordgo.MessageCreate) {
 			db.Where(&Session{UserID: m.Author.ID}).First(&session)
 		}
 
-		if !session.RState { 
+		if !session.RState {
 			// Initiate a session
 			s.ChannelMessageSend(m.ChannelID, "Never Gonna Give u up\nNever gonna let u down\nAlright You may send images now")
 			session.RState = true
 			db.Save(&session)
-		} else { 
+		} else {
 			// Already active session
 			s.ChannelMessageSend(m.ChannelID, "Hold up, you already have an active session\nSend images instead")
 		}
@@ -62,14 +62,14 @@ func PDF(s *discordgo.Session, m *discordgo.MessageCreate) {
 
 		result := db.Where(&Session{UserID: m.Author.ID}).First(&session)
 
-		if result.Error != nil { 
+		if result.Error != nil {
 			// User not in DB
 			s.ChannelMessageSend(m.ChannelID, "I dont even know who u are\nSend soja.start to send me your bank details")
-		} else { 
+		} else {
 			// User in DB
-			if session.RState { 
+			if session.RState {
 				// User in DB + RState True
-				if len(m.Attachments) != 0 { 
+				if len(m.Attachments) != 0 {
 					// soja.f with attachment
 					filePath := session.UserID + "/" + session.UserID + "_" + fmt.Sprint(session.CurrentJPEGs) + ".jpeg"
 					err := dload.DownloadFile(m.Attachments[0].ProxyURL, filePath)
@@ -77,20 +77,20 @@ func PDF(s *discordgo.Session, m *discordgo.MessageCreate) {
 						log.Println("Error downloading file: ", err)
 						return
 					}
-					if isImage(filePath) { 
+					if isImage(filePath) {
 						// File is Image
 						s.ChannelMessageSend(m.ChannelID, "Hippity Hoppty your images are now my property")
 						session.CurrentJPEGs++
 						db.Save(&session)
-					} else { 
+					} else {
 						// Unsupported filetype
 						s.ChannelMessageSend(m.ChannelID, "Error: This file format is not supported")
 					}
-				} else { 
+				} else {
 					// soja.f without attachment
 					s.ChannelMessageSend(m.ChannelID, "Error: no file sent")
 				}
-			} else { 
+			} else {
 				// User in DB + RState False
 				s.ChannelMessageSend(m.ChannelID, "You dont have an active session\nSend soja.start to enable a session")
 			}
@@ -101,52 +101,58 @@ func PDF(s *discordgo.Session, m *discordgo.MessageCreate) {
 	if m.Content == PreCommand+"end" {
 		result := db.Where(&Session{UserID: m.Author.ID}).First(&session)
 
-		if result.Error != nil { 
+		if result.Error != nil {
 			// User doesn't exist/New User
 			s.ChannelMessageSend(m.ChannelID, "I dont even know who u are\nSend soja.start to send me your bank details")
-		} else if session.RState { 
+		} else if session.RState {
 			// User exist + Has an active session
+			if session.CurrentJPEGs != 0 {
+				// Has at least 1 image to convert
 
-			// Reding all images in userID/*.png(jpeg)
-			var inputJPEGs []string
-			for i := 0; i < session.CurrentJPEGs; i++ {
-				inputJPEGs = append(inputJPEGs, session.UserID+"/"+session.UserID+"_"+fmt.Sprint(i)+".jpeg")
+				// Reding all images in userID/*.png(jpeg)
+				var inputJPEGs []string
+				for i := 0; i < session.CurrentJPEGs; i++ {
+					inputJPEGs = append(inputJPEGs, session.UserID+"/"+session.UserID+"_"+fmt.Sprint(i)+".jpeg")
+				}
+
+				// Set metadata for PDFs
+				imp, _ := api.Import("form:A3, pos:c, s:1.0", pdfcpu.POINTS)
+				filePDF := session.UserID + "/" + m.Author.Username + ".pdf"
+
+				// Converting all images in userID/ to userID/userName.pdf
+				err = api.ImportImagesFile(inputJPEGs, filePDF, imp, nil)
+				if err != nil {
+					log.Println("Error Creating output PDF: ", err)
+					return
+				}
+
+				// create *FILE for output.pdf
+				file, err := os.Open(filePDF)
+				if err != nil {
+					log.Println("Error Reading output PDF: ", err)
+					return
+				}
+				defer file.Close()
+
+				// Send the final PDF
+				s.ChannelFileSendWithMessage(m.ChannelID, "Ye Le Bro", m.Author.Username+".pdf", file)
+
+				// Clean all temp directories
+				err = os.RemoveAll(session.UserID)
+				if err != nil {
+					log.Println("Error Removing temp directory: ", err)
+					return
+				}
+
+				// Reset all DB entries except userID
+				session.RState = false
+				session.CurrentJPEGs = 0
+				db.Save(&session)
+			} else {
+				//User exist + Active session + 0 images sent
+				s.ChannelMessageSend(m.ChannelID, "Error: no images found\nSend some images first")
 			}
-
-			// Set metadata for PDFs
-			imp, _ := api.Import("form:A3, pos:c, s:1.0", pdfcpu.POINTS)
-			filePDF := session.UserID + "/" + m.Author.Username + ".pdf"
-
-			// Converting all images in userID/ to userID/userName.pdf
-			err = api.ImportImagesFile(inputJPEGs, filePDF, imp, nil)
-			if err != nil {
-				log.Println("Error Creating output PDF: ", err)
-				return
-			}
-
-			// create *FILE for output.pdf
-			file, err := os.Open(filePDF)
-			if err != nil {
-				log.Println("Error Reading output PDF: ", err)
-				return
-			}
-			defer file.Close()
-			
-			// Send the final PDF 
-			s.ChannelFileSendWithMessage(m.ChannelID, "Ye Le Bro", m.Author.Username + ".pdf", file)
-
-			// Clean all temp directories
-			err = os.RemoveAll(session.UserID)
-			if err != nil {
-				log.Println("Error Removing temp directory: ", err)
-				return
-			}
-
-			// Reset all DB entries except userID
-			session.RState = false
-			session.CurrentJPEGs = 0
-			db.Save(&session)
-		} else { 
+		} else {
 			// User exist + inactive session
 			s.ChannelMessageSend(m.ChannelID, "You don't have any active session\nSend soja.start to initiate a session")
 		}
